backend/example-rag/requirements.txt
openai
pandas
requests
fastapi
uvicorn
faiss-cpu==1.7.4
pydantic
langchain
langchain_openai
python-dotenv
langchain-community